pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "llmtrain"
version = "0.1.0"
description = "Distributed LLM training stack with fault-aware restarts and profiling"
authors = [{name = "Codex", email = "dev@example.com"}]
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
keywords = ["pytorch", "distributed", "deepspeed", "training", "llm"]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "torch>=2.1",
    "torchvision>=0.16",
    "deepspeed>=0.12",
    "pyyaml>=6.0",
    "click>=8.1",
    "tqdm>=4.66",
    "numpy>=1.24",
    "psutil>=5.9",
]

[project.optional-dependencies]
development = ["pytest>=7.4", "mypy>=1.5"]

[project.scripts]
llmtrain = "llmtrain.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}
packages = ["llmtrain"]

[tool.setuptools.package-data]
"llmtrain" = ["py.typed"]

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true
warn_unused_configs = true
warn_return_any = true
warn_unused_ignores = true

